files-view: use inheritance for scripts vars
When a script is present we set some environment variables to allow
the scripts to work with them.

For example, we set a environment variable for selected items.
However, some views have special links, like the desktop. Therefore we
need special treatment for those.

Use inheritance instead of doing special casing in the parent class
for this case.