window-slot: use action state instead of special casing the search
We were special casing to not enable search in some cases where some
special view, like the desktop, doesn't allow it.

But that shouldn't be special cased here, but instead use a subclass of
the window slot and disable it there.

Now that we have that, move the special casing to inheritance for
disabling the search in the desktop.