mime-actions: move opens_in_view to file
And allow overriding of subclasses.

Nautilus mime actions file currently uses the file as parameter, as if
it was part of the nautilus file class.

However it's hosted in a different file, even though we need to special
case some subclasses of nautilus file for its management.
imho, this code design is wrong, since it defeats the purpose of OOP.

So move it to the nautilus file, allow overriding, and override this
vfunc in the subclasses that need it, like the desktop-icon.