general: add uncrustify script
The current style of nautilus is rather poor and mixes at least 3
different code styles.

Specific issues that affect my personal performance are:
- tabs for multiline alignment.
- mix of tabs and spaces.
- errors on no braced one liners conditionals.
- errors on non braced case statements with variable declarations.

So I would say requirements for the style is to address the previous
issues and also be well known. I don't want new contributors to see a
new style completely different from C books authors.

So far, I found Allman (aka BSD) style which seems the choice of most C
books authors as far as I can see on internet and address the previous
issues.

So as a matter of practicity I'm going to convert all Nautilus style to
this one, and if the next person who contributes most on Nautilus has
a different choice, please feel free to change whatever makes your
performance the best.