nautilus-desktop: Place icons only in primary monitor
Due to the restrictions of using a big window we were making choices
between different areas across monitors for placing icons.
However that makes happy no one since depending on the area used the
icons can jump or not being places on the primary monitor on certain
multi monitor setups.

For now use a single window that equals to the primary monitor to avoid
these issues.
For future we might want to have a single window per each monitor.