Add sorting options for files names
Now the files names can be sorted by using criteria like: original name or
last modified date. This is useful when there is used a numbering option and
the user is not pleased with the default order of the files and wants to
change it.

In this commit some warnigns were also fixed.